compaction_manager: pass std::string_view when appropriate
most of the times, we just pass a string literal for the `op` argument,
and we just print it in the logging message, so there is no need to
create an instance of sstring for formatting it.

in this change, we use std::string_view for the type of this parameter
for better readability.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>