compaction_manager: swallow gate_closed_exception if stopping
we can perform cleanup compaction on a table, and the table is
being dropped at the same time. under the hood, the cleanup
operation can be executed as multiple coroutines to compact
the different compaction group owned by the table, we do try
to abort the ongoing compaction job of the table. but still,
when calling into `compaction_manager::try_perform_cleanup()`,
this function tries to hold the compaction_state.gate of the
table. if the `table::stop()` closes gate with
`compaction_group:stop()` first,
`compaction_manager::try_perform_cleanup()` throws a
`gate_closed_exception` which gets populated to the outer
caller of `run_on_table()` which log this down with ERROR log
level, so the test or human developer would notice this and
consider it an indication of a bug.

but it is not.

in this change, we just catch the exception and check if the
table is being stopped, only rethrow it if we are not stopping.

Fixes #16305
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>