docs: 5.2->5.4 upgrade: add 'Validate Raft setup' section
A similar section appears in the 5.1->5.2 upgrade procedure, we also
need it in 5.2->5.4.

These sections will have to be included in all upgrade guides until the
version we're upgrading *from* is enabling Raft unconditionally. First
such planned version is 5.5; therefore (if everything goes according to
plan) the 5.5->later upgrade guide will no longer need the section, but
5.2->5.4 and 5.4->5.5 and corresponding enterprise versions still need
it.

This commit needs to be backported to 5.4 and corresponding enterprise
release branch.

Refs: scylladb/scylladb#15854, scylladb/scylladb#16334